refactor(l1, l2): abstract common logic (#2109)
